requirements.txt
python-telegram-bot==12.7
discord-webhook==0.16.3
selenium==4.3.0
webdriver-manager==3.8.0